feat: cc-0101 remove courtName and courtSizeSlice
