+++ src/components/BasketballCourt/DashedLine.tsx
@@ -1,6 +1,6 @@
 import { Shape } from "react-konva";
 import { useStoreSelector } from "@/store/hooks";
-import { dashedWhiteLine } from "../../store/reducer/courtSizeSlice";
+import { dashedWhiteLine } from "../../store/reducer/courtSpecDataSlice";
 import { ICourtStartPoint } from "@/interfaces/courtStartPoint";
 
 interface DashedLineProps {
