+++ src/store/index.ts
@@ -1,6 +1,4 @@
 import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit";
-import courtReducer from "./reducer/courtSizeSlice";
-import courtNameReducer from "./reducer/courtNameSlice";
 import userReducer from "./reducer/userSlice";
 import tileReducer from "./reducer/tileSlice";
 import courtColorReducer from "./reducer/courtColorSlice";
@@ -16,8 +14,6 @@ import courtSpecDataReducer from "./reducer/courtSpecDataSlice";
 export const makeStore = () =>
   configureStore({
     reducer: {
-      // courtSize: courtReducer,
-      // courtName: courtNameReducer,
       tile: tileReducer,
       priceBar: priceBarReducer,
       courtColor: courtColorReducer,
